backend/pyproject.toml
[project]
name = "task-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
